+++ waterfun-service/src/main/java/org/waterwood/waterfunservice/service/post/TagService.java
@@ -23,7 +23,7 @@ public interface TagService {
      * @param id tag id
      * @return {@link Tag}
      */
-    Tag getTag(Integer id);
+    Tag getTag(Long id);
 
     /**
      * Update a tag
@@ -35,14 +35,14 @@ public interface TagService {
      * Delete a tag
      * @param id tag id
      */
-    void deleteTag(Integer id);
+    void deleteTag(Long id);
 
     /**
      * Get tags by tag ids
      * @param tagIds tag ids
      * @return list ofPending {@link Tag}
      */
-    Set<Tag> getTags(Iterable<Integer> tagIds, boolean strict);
+    Set<Tag> getTags(Iterable<Long> tagIds, boolean strict);
 
     /**
      * Create new tags by names
